fix(s2d): honest sourcing for single-parity vector; share PlatformVector via harness
