+++ .planning/phases/18-quality-audit/18-AUDIT.md
@@ -15,7 +15,7 @@ Newly audited here: S2D, Nutanix, NetApp, Ceph, Synology, Longhorn + PPTX export
 
 | # | Platform/Area | Tag | Severity | Description | Reference (source + URL) | Status |
 |---|---------------|-----|----------|-------------|--------------------------|--------|
-| 1 | S2D | untested | — | No external-reference vector coverage before phase 18. Added 4 vectors (3-way mirror, single parity, dual parity @7 FDs hybrid, mirror-accelerated parity @7 FDs). All pass at 0.00% deviation — no engine change needed. | Microsoft Learn plan-volumes / fault-tolerance / mirror-accelerated-parity (URLs in Reference Cases → S2D) | untested → now covered |
+| 1 | S2D | untested | — | No external-reference vector coverage before phase 18. Added 4 vectors (3-way mirror, single parity [engine-formula analog — no MS-published fraction exists], dual parity @7 FDs hybrid, mirror-accelerated parity @7 FDs). All pass at 0.00% deviation — no engine change needed. | Microsoft Learn plan-volumes / fault-tolerance / mirror-accelerated-parity (URLs in Reference Cases → S2D) | untested → now covered |
 
 Tags: value-wrong (>1% off reference) · value-misleading (right number, wrong label/unit) · untested (no vector coverage)
 
@@ -45,7 +45,7 @@ Drive: `testDrive1TB` (HDD ⇒ engine selects the *hybrid* dual-parity stepped t
 | Config | MS efficiency fraction | Source | Expected usable (bytes) | Engine (bytes) | Deviation |
 |--------|-----------------------|--------|------------------------|----------------|-----------|
 | 3-way mirror, 12×1 TB, 4 servers (mirrorCopies=3) | 33.3% (1/3) | [Plan volumes — mirror efficiency](https://learn.microsoft.com/en-us/windows-server/storage/storage-spaces/plan-volumes) | 2 341 873 333 333 | 2 341 873 333 333 | 0.00% |
-| Single parity, 16×1 TB, 4 servers (faultDomains=4) | 75% ((N−1)/N) | [Plan volumes — parity efficiency](https://learn.microsoft.com/en-us/windows-server/storage/storage-spaces/plan-volumes) | 8 548 540 000 000 | 8 548 540 000 000 | 0.00% |
+| Single parity, 16×1 TB, 4 servers (faultDomains=4) — **engine-formula analog, see honesty note** | 75% ((N−1)/N, RAID-5 analogy) | [Fault tolerance — single parity (qualitative only)](https://learn.microsoft.com/en-us/windows-server/storage/storage-spaces/fault-tolerance) | 8 548 540 000 000 | 8 548 540 000 000 | 0.00% |
 | Dual parity, 16×1 TB, 7 servers (faultDomains=7, hybrid) | 66.7% (RS 4+2, 7–11 FDs) | [Fault tolerance & storage efficiency](https://learn.microsoft.com/en-us/windows-server/storage/storage-spaces/fault-tolerance) | 7 568 540 000 000 | 7 568 540 000 000 | 0.00% |
 | Mirror-accelerated parity, 16×1 TB, 7 servers (20/80 tiering, mirrorCopies=2) | 63.3% (0.2/2 + 0.8×2/3) | [Mirror-accelerated parity (ReFS)](https://learn.microsoft.com/en-us/windows-server/storage/refs/mirror-accelerated-parity) | 7 176 540 000 000 | 7 176 540 000 000 | 0.00% |
 
@@ -56,6 +56,14 @@ Note: the MAP 20/80 mirror:parity split is Microsoft's *typical* configuration (
 tunable per volume); the vector validates the engine's documented 20/80 model, not a fixed
 Microsoft constant. The 277 GB infra reserve and rebuild-reserve sizing are engine policy
 sourced in code comments (Azure Local docs); this task validated the resiliency fractions.
+
+**Honesty note (single parity):** Microsoft documents single parity only qualitatively on the
+fault-tolerance page ("keeps only one bitwise parity symbol … most closely resembles RAID-5")
+and publishes **no numeric single-parity efficiency fraction** anywhere on Learn (re-verified
+via Perplexity, 2026-07-11). The (N−1)/N value is the standard RAID-5 analogy — the same
+formula the engine implements — so this vector is an *engine-formula analog* (regression pin),
+not an independent external validation. Externally validated vectors: 3/4 (3-way mirror,
+dual parity, MAP); coverage should not be overstated as 4/4 external.
 
 ## Spot-Checks (Task 9)
